+++ backend/lm_backend/api/routes/bookings.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.booking import BookingCRUD
-from lm_backend.api.models import Booking
+from lm_backend.api.models.booking import Booking
 from lm_backend.api.schemas import BookingCreateSchema, BookingSchema, BookingUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
