+++ backend/lm_backend/api/routes/configurations.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.generic import GenericCRUD
-from lm_backend.api.models import Configuration
+from lm_backend.api.models.configuration import Configuration
 from lm_backend.api.schemas import ConfigurationCreateSchema, ConfigurationSchema, ConfigurationUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
