+++ backend/lm_backend/api/routes/inventories.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.generic import GenericCRUD
-from lm_backend.api.models import Inventory
+from lm_backend.api.models.inventory import Inventory
 from lm_backend.api.schemas import InventoryCreateSchema, InventorySchema, InventoryUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
