+++ backend/lm_backend/api/routes/jobs.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.generic import GenericCRUD
-from lm_backend.api.models import Job
+from lm_backend.api.models.job import Job
 from lm_backend.api.schemas import JobCreateSchema, JobSchema, JobUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
