+++ backend/lm_backend/api/routes/license_servers.py
@@ -3,8 +3,9 @@ from typing import List, Optional
 from fastapi import APIRouter, Depends, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from lm_backend.api.constants import LicenseServerType
 from lm_backend.api.cruds.generic import GenericCRUD
-from lm_backend.api.models import LicenseServer
+from lm_backend.api.models.license_server import LicenseServer
 from lm_backend.api.schemas import LicenseServerCreateSchema, LicenseServerSchema, LicenseServerUpdateSchema
 from lm_backend.api.constants import LicenseServerType
 from lm_backend.permissions import Permissions
