+++ backend/tests/test_database.py
@@ -9,7 +9,7 @@ from pytest import fixture, raises
 from sqlalchemy import select
 
 from lm_backend import database
-from lm_backend.api.models import Product
+from lm_backend.api.models.product import Product
 
 _LIST_TABLES_SQL = "SELECT name FROM sqlite_master WHERE type='table' AND name NOT LIKE 'sqlite_%'"
